# server/swagger.json
{
    "securityDefinitions": {
      "jwt": {
          "type": "apiKey",
          "in": "header",
          "name": "Authorization"
      }
    },
    "security": [
        {
            "jwt": []
        }
    ],
    "swagger": "2.0",
    "info": {
      "version": "1.0.0",
      "title": "Pets Gallery",
      "description": "Team tak21tiim2 project",
      "license": {
        "name": "MIT",
        "url": "https://opensource.org/licenses/MIT"
      }
    },
    "host": "localhost:4111",
    "basePath": "/",
...
    "schemes": ["http"],
    "consumes": ["application/json"],
    "produces": ["application/json"],
    "paths": {
...
      "/api/get-pets-data/{userId}": {
        "post": {
          "tags": ["Pets"],
          "description": "Get all user's pets in system database",
          "parameters": [
            {
              "in": "path",
              "name": "userId",
              "type": "string",
              "example": "Urmas",
              "required": true
            },
            {
              "in": "body",
              "name": "page",
              "type": "array",
              "example": "{\"page\":2}",
              "default": "{\"page\":1}",
              "description": "One pet per page - change page number to get another pet data",
              "required": false
            }
          ],
          "produces": ["application/json"],
          "responses": {
            "200": {
              "description": "OK",
              "schema": {
                "$ref": "#/definitions/Pets"
              }
            }
          }        
        }    
      },
...
      "/api/edit-pet": {
        "post": {
          "tags": ["Pets"],
          "description": "Edit user's pet",
          "parameters": [
            {
              "in": "body",
              "name": "dataArray",
              "type": "array",
              "example": "{\"pet_id\": \"62188293093e4c6e3ed7e941\",\"omanik\": \"Urmas\",\"loomaNimi\": \"TestikasXXX\",\"liik\": \"Liik\",\"vanus\": 5,\"sugu\": \"Emane\",\"v2limus\": \"sadsadsad\",\"iseloom\": \"sadsadsadsa\"}",
              "required": true
            }
          ],
          "produces": ["application/json"],
          "responses": {
            "200": {
              "description": "OK",
              "schema": {
                "$ref": "#/definitions/EditPet"
              }
            },
            "401": {
              "description": "Unauthorized",
              "schema": {
                "$ref": "#/definitions/EditPet401"
              }
            }
          }        
        }    
      },
      "/api/delete-pets/{id}": {
        "get": {
          "tags": ["Pets"],
          "description": "Delete user's pet from system database",
          "parameters": [
            {
              "in": "path",
              "name": "id",
              "type": "string",
              "example": "619f2cdbede8a2251b4f49e6",
              "required": true
            }
          ],
          "components": {
                "securitySchemes": {
                  "jwt": {
                    "type": "http",
                    "scheme": "bearer",
                    "in": "header",
                    "bearerFormat": "JWT"
                  }
                }
              },
              "security": [{
                "jwt": []
              }],
          "produces": ["application/json"],
          "responses": {
            "200": {
              "description": "OK"
            }
          }        
        }    
      },
      "/api/add-comment/{id}": {
        "post": {
          "tags": ["Pets"],
          "description": "Add comment to user's pet",
          "parameters": [
            {
              "in": "path",
              "name": "id",
              "type": "string",
              "example": "62188293093e4c6e3ed7e941",
              "required": true
            },
            {
              "in": "body",
              "name": "dataArray",
              "type": "array",
              "example": "{\"message\": \"Message\",\"username\": \"Urmas\"}",
              "required": true
            }
          ],
          "produces": ["application/json"],
          "responses": {
            "200": {
              "description": "OK",
              "schema": {
                "$ref": "#/definitions/EditPet"
              }
            },
            "401": {
              "description": "Unauthorized",
              "schema": {
                "$ref": "#/definitions/EditPet401"
              }
            }
          }        
        }    
      }
    },
    "definitions": {
...
      "Pets": {
        "required": ["docs", "totalDocs", "limit", "totalPages", "page", "sugu", "pagingCounter", "hasPrevPage", "hasNextPage", "prevPage", "nextPage"],
        "properties": {
          "docs": {
            "type": "array",
            "items": {
              "type": "object",
              "properties": {
                "_id": {
                  "type": "string"
                },
                "omanik": {
                  "type": "string"
                },
                "loomaNimi": {
                  "type": "string"
                },
                "liik": {
                  "type": "string"
                },
                "vanus": {
                  "type": "number"
                },
                "sugu": {
                  "type": "string"
                },
                "v2limus": {
                  "type": "string"
                },
                "iseloom": {
                  "type": "string"
                },
                "pilt": {
                  "type": "string"
                },
                "pilt64_s": {
                  "type": "string"
                },
                "kommentaarid": {
                  "type": "array",
                  "items": {
                    "type": "object",
                    "properties": {
                "username": {
                  "type": "string"
                },
                "message": {
                  "type": "string"
                },
                "_id": {
                  "type": "string"
                }
              }
                        }
                }
              }
            }
          },
          "totalDocs": {
            "type": "number"
          },
          "limit": {
            "type": "number"
          },
          "totalPages": {
            "type": "number"
          },
          "page": {
            "type": "number"
          },
          "pagingCounter": {
            "type": "number"
          },
          "hasPrevPage": {
            "type": "boolean"
          },
          "hasNextPage": {
            "type": "boolean"
          },
          "prevPage": {
            "type": "number"
          },
          "nextPage": {
            "type": "number"
          }
        }
      },
...
      "EditPet": {
        "required": ["pet_id","omanik", "loomaNimi", "liik", "vanus", "sugu", "v2limus", "iseloom"],
        "type": "object",
        "properties": {
          "status": {
            "type": "string"
          }
        }
      },
      "EditPet401": {
        "required": ["pet_id","omanik", "loomaNimi", "liik", "vanus", "sugu", "v2limus", "iseloom"],
        "type": "object",
        "properties": {
          "status": {
            "type": "string"
          }
        }
      }
    }
  }